+++ R/fuj-package.R
@@ -5,8 +5,23 @@
 ## usethis namespace: end
 NULL
 
+#' `{fuj}` options
+#'
+#' Options uses for `{fuj}` functions.
+#'
+#' @keywords internal
+#' @examples
+#' names(op.fuj)
+#' op.fuj
+#' @export
+# nolint next: object_name_linter.
 op.fuj <- list(
   fuj.verbose = NULL,
   fuj.verbose.fill = FALSE,
-  fuj.verbose.label = "[verbose] "
+  fuj.verbose.label = "<verboseMessage> "
 )
+
+.onLoad <- function(libname, pkgname) {
+  # set options
+  options(op.fuj[!names(op.fuj) %in% names(options())])
+}